Resolve mr.hosseinalipour comments.
