+++ BSN.Commons.sln
@@ -43,12 +43,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BSN.Commons.Orm.EntityFrame
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BSN.Commons.Users", "Source\BSN.Commons.Users\BSN.Commons.Users.csproj", "{213ABCEF-7E9A-4CE5-A3EF-289C9781344D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BSN.Commons.S3.Abstraction", "Source\BSN.Commons.S3.Abstraction\BSN.Commons.S3.Abstraction.csproj", "{F9D348C7-F468-4F06-A0FB-3B9C6116555E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BSN.Commons.S3.Core", "Source\BSN.Commons.S3.Core\BSN.Commons.S3.Core.csproj", "{48F4694B-2732-472B-A05E-4B2926B3B866}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BSN.Commons.S3.Net", "Source\BSN.Commons.S3.Net\BSN.Commons.S3.Net.csproj", "{5F798AA1-1A65-4C9F-8415-3100C141DD87}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -87,18 +81,6 @@ Global
 		{213ABCEF-7E9A-4CE5-A3EF-289C9781344D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{213ABCEF-7E9A-4CE5-A3EF-289C9781344D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{213ABCEF-7E9A-4CE5-A3EF-289C9781344D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F9D348C7-F468-4F06-A0FB-3B9C6116555E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F9D348C7-F468-4F06-A0FB-3B9C6116555E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F9D348C7-F468-4F06-A0FB-3B9C6116555E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F9D348C7-F468-4F06-A0FB-3B9C6116555E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{48F4694B-2732-472B-A05E-4B2926B3B866}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{48F4694B-2732-472B-A05E-4B2926B3B866}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{48F4694B-2732-472B-A05E-4B2926B3B866}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{48F4694B-2732-472B-A05E-4B2926B3B866}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5F798AA1-1A65-4C9F-8415-3100C141DD87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5F798AA1-1A65-4C9F-8415-3100C141DD87}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5F798AA1-1A65-4C9F-8415-3100C141DD87}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5F798AA1-1A65-4C9F-8415-3100C141DD87}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -112,9 +94,6 @@ Global
 		{906FEED8-23E0-4EEF-B902-C39325C50480} = {5C6BA7B5-832A-495A-AF5E-C2A74F6A1EF9}
 		{335F645B-C85F-42C2-9185-A216101F60C7} = {DC377ADC-CC9D-4785-81BE-726DBF5F3096}
 		{213ABCEF-7E9A-4CE5-A3EF-289C9781344D} = {DC377ADC-CC9D-4785-81BE-726DBF5F3096}
-		{F9D348C7-F468-4F06-A0FB-3B9C6116555E} = {DC377ADC-CC9D-4785-81BE-726DBF5F3096}
-		{48F4694B-2732-472B-A05E-4B2926B3B866} = {DC377ADC-CC9D-4785-81BE-726DBF5F3096}
-		{5F798AA1-1A65-4C9F-8415-3100C141DD87} = {DC377ADC-CC9D-4785-81BE-726DBF5F3096}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BCAF76D3-AA3C-4D0F-8D10-34065F8FED09}
