--- Source/BSN.Commons.S3.Abstraction/BSN.Commons.S3.Abstraction.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-	<LangVersion>10.0</LangVersion>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\BSN.Commons\BSN.Commons.csproj" />
-  </ItemGroup>
-
-</Project>
