--- Source/BSN.Commons.S3.Core/BSN.Commons.S3.Core.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-	  <TargetFrameworks>net6.0</TargetFrameworks>
-      <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="AWSSDK.S3" Version="3.7.305.2" />
-    <PackageReference Include="Microsoft.Extensions.Options" Version="6.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\BSN.Commons.S3.Abstraction\BSN.Commons.S3.Abstraction.csproj" />
-  </ItemGroup>
-
-</Project>
