--- Source/BSN.Commons.S3.Net/packages.config
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="AWSSDK.Core" version="3.7.300.27" targetFramework="net472" />
-  <package id="AWSSDK.S3" version="3.7.305.2" targetFramework="net472" />
-  <package id="AWSSDK.SecurityToken" version="3.7.300.28" targetFramework="net472" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="6.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="6.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Options" version="6.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Primitives" version="6.0.0" targetFramework="net472" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.4" targetFramework="net472" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net472" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" />
-</packages>
