+++ Source/BSN.Commons/BSN.Commons.csproj
@@ -55,6 +55,7 @@
   <ItemGroup>
     <PackageReference Include="Apache.NMS" Version="2.0.0" />
     <PackageReference Include="Apache.NMS.ActiveMQ" Version="2.0.0" />
+    <PackageReference Include="AWSSDK.S3" Version="3.7.305.2" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.32" />
     <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.32" />
     <PackageReference Include="Confluent.Kafka" Version="2.3.0" />
